# eims-project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ExamFlow EIMS API
  description: |
    Exam Information Management System API for UP Tacloban.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ## Rate Limiting
    - 100 requests per 15 minutes per IP
    - Rate limit headers included in responses
    
    ## WebSocket
    Real-time updates available at `ws://localhost:3001/ws?token=<access_token>`
  version: 1.0.0
  contact:
    name: ExamFlow Support
    email: support@examflow.up.edu.ph
  license:
    name: Educational Use
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://examflow.up.edu.ph/api/v1
    description: Production server
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: faculty@up.edu.ph
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: integer
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logout successful
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
...
  /announcements:
    get:
      tags: [Announcements]
      summary: List announcements
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AnnouncementType'
      responses:
        '200':
          description: List of announcements
    post:
      tags: [Announcements]
      summary: Create announcement
      description: Create announcement and trigger real-time notifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnnouncementRequest'
      responses:
        '201':
          description: Announcement created and notifications sent
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
    Role:
      type: string
      enum: [ADMIN, FACULTY, STUDENT]
...
    AnnouncementType:
      type: string
      enum: [GENERAL, EXAM_UPDATE, SCHEDULE_CHANGE, RESULT_RELEASE, PROCTOR_NOTICE, EMERGENCY]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
...
    CreateAnnouncementRequest:
      type: object
      required: [title, content, type, targetRoles]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        content:
          type: string
          minLength: 10
        type:
          $ref: '#/components/schemas/AnnouncementType'
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
        examId:
          type: string
          format: uuid
        targetRoles:
          type: string
          description: Comma-separated roles
          example: "STUDENT,FACULTY"
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              type: object
              properties:
                accessToken:
                  type: string
                refreshToken:
                  type: string
                expiresIn:
                  type: integer
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...